feat(functions.zsh): add new function cp_docker to copy files between Docker containers and the host machine
The new function cp_docker is added to the zsh/common/functions.zsh file. This function uses the 'docker run' command to run a temporary Alpine container and copies files from the source directory to the destination directory using the 'cp' command. This function is useful for easily copying files between Docker containers and the host machine.